pyproject.toml
[project]
name = "wavegc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    # Core scientific stack
    "numpy>=2.0",
    "scipy>1.10",
    "pandas",
    "matplotlib",
    "seaborn",
    "plotly>=6.4.0",
    "statsmodels",
    "scikit-learn>=1.6.1",
    # PyTorch, PyG, ...
    # To Do: add gpu support
    "torch>=2.0",
    "torch_geometric>=2.3.0",
    # Notebooks, loggers & visualization
    "ipywidgets>=8.1.7",
    "ipykernel",
    "nbformat>=5.10.4",
    "wandb",
    # Progress bars & misc
    "tqdm>=4.67.1",
    "pip",
    # clean code & experementation
    "hydra-core",
    "black",
    "isort",
    "pre-commit",
    "ogb>=1.3.6",
    "torch-sparse>=0.6.18",
    "torch-scatter>=2.1.2",
    "torchmetrics>=1.8.2",
]
